curves reorg
